Fix issue #1444 Pressure plates activate with creativeNoClip if another entity is in the same block
I added a mixin to intercept isIgnoringBlockTriggers() in getSignalStrength when the entity is the player in creativeNoClip mode and flying. We just report that we are ignoring triggers.

Verified that issue happens without this mixin, and goes away with it (the pressure plate no longer fires when the user flies through the plate).