add zeroTickPlants rule
"Restores the ability to force plants (Bamboo [NOT SHOOTS], Cactus, Chorus Flower, and Sugar Cane) to advance their growth stage with zero-ticking"
I'm not sure what carpets policy on @ Overwrite is, but in this case I believe it's a better approach than @ Inject, since we need the else branch in order to ensure destroyBlock() and randomTick() remain mutually exclusive. This ensures that growing plants with zero-ticking will work exactly as it did before being patched, since this mirrors the vanilla implementation that caused the behavior.